feat: Refactor videoSearch component and separate API function
- Moved YouTube API call logic from videoSearch.jsx to a new utils/youtubeApi.js file.
- Updated videoSearch.jsx to import and use the getVideos function from youtubeApi.js.
- Ensured code modularity and improved organization by separating concerns.
- Fixed ESLint warning related to Fast Refresh by keeping only the component in videoSearch.jsx.
